Updating .clang-tidy and move sse2neon.h into its own helper subdir - not sure it'll definitely stay there, yet.
