+++ src/crispy/StrongLRUHashtable.h
@@ -25,7 +25,7 @@
 #if defined(__x86_64__)
     #include <immintrin.h>
 #elif defined(__aarch64__)
-    #include <crispy/sse2neon.h>
+    #include <sse2neon/sse2neon.h>
 #endif
 
 #define DEBUG_STRONG_LRU_HASHTABLE 1
