+++ src/sse2neon/README.md
@@ -1,0 +1,3 @@
+This is an external depenency that used to live within the crispy library.
+
+This is now moved out in order to make clang-tidy a little bit more happy and not cry about external code.
